requirements.txt
streamlit==1.33.0
langchain==0.1.13
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.38.2
torch==2.1.2
faiss-cpu==1.7.4
pdfminer.six==20221105
python-dotenv==1.0.1
uvicorn==0.27.1
requests==2.31.0
fastapi==0.110.0